+++ Scripts/Java/basex.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-java -jar $HOME/Documents/Estudios/DAW/LND/Tema6/BaseX104.jar