+++ Scripts/Java/saxon.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-java -jar $HOME/Documents/Estudios/DAW/LND/Tema5/saxon-he-11.4/saxon-he-11.4.jar